PMM-6826 Refactoring.
